Основной сервис

+++ main-service/src/main/java/ru/practicum/event/controller/EventControllerAdmin.java
@@ -29,7 +29,7 @@ public class EventControllerAdmin {
     @GetMapping
     public List<EventRespShort> get(@RequestParam(required = false) List<Long> users,
                                    @RequestParam(required = false) List<String> states,
-                                   @RequestParam(required = false) List<Long> categories,
+                                   @RequestParam(required = false) List<Integer> categories,
                                    @RequestParam(required = false) @DateTimeFormat(
                                            pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeStart,
                                    @RequestParam(required = false) @DateTimeFormat(
